Collect all unused custom horizontal calendar code together
